Large Update
This is a large commit, containg several updates to the program.

Updates to the application to support the Z-stage AutoFocus, including:
* Building a Z-Stacked Image.
* allowing for better control of the lighting during image generation
(though more work is needed)
* Re-designing the ComPort creation to allow for it to be handeled all
inside of appDesigner
* Laying the groud work for much more features to be added quickly,
It still needs a way to remember the setting the user gave it from the
last time it ran, so that you dont have to add it everytime the program
starts up.

A lot of other utility function have been added to the program, some of
which may be able to be installed as a dependant library in the future,
Most noteably the *+fig* folder which what found on [Matlab Answers]
(https://www.mathworks.com/matlabcentral/answers/98331-is-it-possible-to-maximize-minimize-or-get-the-state-of-my-figure-programmatically-in-matlab)
Though this is only a work around for a function comming out next year.